Use longs instead of integer

+++ model-events/src/main/java/de/focus_shift/urlaubsverwaltung/extension/api/application/VacationTypeDTO.java
@@ -9,7 +9,7 @@ import lombok.extern.jackson.Jacksonized;
 @Builder
 @Jacksonized
 public class VacationTypeDTO {
-  @NonNull Integer sourceId;
+  @NonNull Long sourceId;
   @NonNull String category;
   boolean requiresApprovalToApply;
   boolean requiresApprovalToCancel;
