+++ model-events/src/main/java/de/focus_shift/urlaubsverwaltung/extension/api/sicknote/SickNoteCancelledEventDTO.java
@@ -15,7 +15,7 @@ import lombok.extern.jackson.Jacksonized;
 public class SickNoteCancelledEventDTO {
   @NonNull private UUID id;
 
-  @NonNull private Integer sourceId;
+  @NonNull private Long sourceId;
 
   @NonNull private Instant createdAt;
